refactor: share file download helper
